Fix intermittent slowness in resolver activity towards end of day
The resolver calls UsageStatsManager.queryUsageStats() to get package usage information.
This causes all XMLs of the selected bucket to be read, which can be of significant size.
However, the majority of that size is due to usage event information, which queryUsageStats()
isn't interested in, so it'll parse a lot of data just to throw it away later on.

Solution: Don't parse the portions of the XMLs we aren't interested in.

Change-Id: Ie41207a3ce2dc5b7822549666b6ed0e10e339e92
BenzoEdit: Fixed UserUsageStatsService for Pie